Add files via upload
I had added a Scrolbar, delete task button and some other minior features such as making sure when the program is launched it will be in the center of the users screen. I did this because I find it to be most convienient. I also changed the way some of my things are named because it was knd of confusing before. I might of missed some but I will have to go through it in my version 3 when im doing my testing.