Add files via upload
This is the final versioin of my program. Its found in my N drive. Finished on 27/08/2025.
he majority of this code belongs to me with the accetion of mabye 7 lines. These pieces of code is marked in my previus version. They are not marked in this version as i think it would be not needed and would take up space